Adds a bit of clarity in the side effects object itself...
Not sure about whether it makes the use of it more clear
